Add new method to A* class
This new method is used to add a node to the search queue. This method
calculated the Manhattan distance for every child of the, current, node
and only adds those which calculated Manhattan distance is lower than
the original calculated one to the search queue.
JavaDoc for this method has been written.

Testing shows that the algorithm performs perfectly for the first 2 test
scenarios, where the estimated path cost and the actual path cost
matches. Only on the 3 test scenario does it underestimate the actual
path cost